Refactor Scoring into a Score model
Prior the scoring code was all a utility function that always took in
and scored a hand. That code has been split into 2 locations. The new
BasicScore class provides basic scoring functionality for a hand of
crib, and the new score_helper provides some basic utility functions.

This was done to more closely model the domain.